autoApiTest_shop/requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
Appium-Python-Client==0.11
arrow==1.2.2
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
basepage==1.7.4
blinker==1.4
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
click-aliases==1.0.1
click-default-group==1.2.2
click-help-colors==0.9.1
colorama==0.4.4
commonmark==0.9.1
ConfigArgParse==1.5.3
cryptography==36.0.2
cssselect==1.1.0
custom-text-test-runner==0.1.5
deepdiff==5.5.0
enum34==1.1.10
et-xmlfile==1.1.0
execnet==1.9.0
fake-useragent==0.1.11
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
futures==3.0.5
genson==1.2.2
gevent==21.12.0
geventhttpclient==2.0
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==2.7
iniconfig==1.1.1
itsdangerous==2.1.2
javaobj-py3==0.4.3
Jinja2==3.1.1
jmespath==0.10.0
jsonmerge==1.8.0
jsonpath==0.82
jsonschema==4.4.0
kaitaistruct==0.9
ldap3==2.9.1
locust==2.10.1
lxml==4.8.0
MarkupSafe==2.1.1
mitmproxy==8.0.0
msgpack==1.0.3
networkx==2.8.4
numpy==1.20.3
openpyxl==3.0.10
ordered-set==4.0.2
outcome==1.2.0
packaging==20.9
pandas==1.5.3
parse==1.19.0
passlib==1.7.4
Pillow==9.2.0
pluggy==0.13.1
protobuf==3.19.4
psutil==5.9.1
publicsuffix2==2.20191221
py==1.10.0
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
Pygments==2.13.0
PyHamcrest==2.0.3
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
pystache==0.6.0
pytest==6.2.5
pytest-assume==2.4.3
pytest-forked==1.4.0
pytest-rerunfailures==9.1.1
pytest-sugar==0.9.4
pytest-xdist==2.5.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
pyzmq==22.3.0
redis==4.4.2
requests==2.28.1
requests-xml==0.2.3
rich==9.13.0
roundrobin==0.0.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.1
schemabuilder==0.3.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.40
table-printer==0.1.4
tabulate==0.8.10
task==0.2.5
termcolor==1.1.0
toml==0.10.2
tornado==6.1
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.12
urwid==2.1.2
w3lib==1.22.0
Werkzeug==2.0.3
wsproto==1.1.0
xlrd==2.0.1
XMind==1.2.0
xmind2testcase==1.5.0
xmljson==0.2.1
zope.event==4.5.0
zope.interface==5.4.0
zstandard==0.17.0
